Refactor: 로그아웃 로직을 유틸 함수로 분리하여 재사용성 및 가독성 향상
